to fix Lesson 13 task4

+++ src/lesson13/task4/Calkulator.java
@@ -6,18 +6,18 @@ package lesson13.task4;
  */
 public class Calkulator {
     public static <T extends Number, V extends Number> double sum(T t, V v) {
-        return (int) t.doubleValue() + v.doubleValue();
+        return  t.doubleValue() + v.doubleValue();
     }
 
     public static <T extends Number, V extends Number> double minus(T t, V v) {
-        return (int) t.doubleValue() - v.doubleValue();
+        return t.doubleValue() - v.doubleValue();
     }
 
     public static <T extends Number, V extends Number> double generation(T t, V v) {
-        return (int) t.doubleValue() * v.doubleValue();
+        return  t.doubleValue() * v.doubleValue();
     }
 
     public static <T extends Number, V extends Number> double division(T t, V v) {
-        return (int) t.doubleValue() / v.doubleValue();
+        return t.doubleValue() / v.doubleValue();
     }
 }
